requirements.txt
Django==4.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
django-cors-headers==4.3.1
pytest==7.4.3
pytest-django==4.7.0
black==23.11.0
flake8==6.1.0
django-cors-headers==4.3.1